Created a .blend file to represent an empty IIIF Manifest. This could be used as a template for user wishing to generate a new manifest by loading this file as the initial Blender scene, adding models, cameras and lights to the AnnotationPage collection, then exporting the scene as an IIIF manifest.
This .blend file also contains a Python script to be run after the scene is loaded. The script can be viewed and edited in the Text Editor panel, in the log_open.py text block. All this script does is print a message to console, it could late be used to generate additional feature of the manifest template